requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0